moved scheduled learning days and teams learned topics
